% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddies_matrix.R
\name{buddies_matrix}
\alias{buddies_matrix}
\title{Creating Buddies Matrices}
\usage{
buddies_matrix(
  data = data,
  k = 3,
  subsample_prop = 0.8,
  num_subsamples = 10,
  start_seed = 123,
  cluster_method
)
}
\arguments{
\item{data}{The dataset to be analyzed}

\item{k}{The specific number of clusters (k) to evaluate. Defaults to 3.}

\item{subsample_prop}{Proportion of the dataset to use for each resample. Defaults to 0.8.}

\item{num_subsamples}{Number of resamples to generate. Defaults to 10.}

\item{start_seed}{Random seed for reproducibility. Defaults to 123.}

\item{cluster_method}{Specified clustering algorithm (i.e. "kmeans")}
}
\value{
Produces a list of matrices, where the number of matrices
corresponds to num_subsamples.
}
\description{
Generates buddies matrix for each resample corresponding to a specified k.
The number of matrices produces is equivalent to the number of subsamples. In
these matrices, rivals are assigned a value of -1 and buddies are assigned a
value of 1. Rivals belong to different clusters, while buddies belong to the
same cluster. The function applies the k-means algorithm via the k_means()
function.
}
\examples{
data <- simulate_mvn(n = 10)
buddy <- buddies_matrix(data = data, cluster_method = "kmeans")
buddy[[1]]
buddy[[2]]
}
