% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crab.R
\name{crab}
\alias{crab}
\title{Running the CRAB Algorithm}
\usage{
crab(
  data,
  n = 100,
  min_k = 2,
  max_k = 5,
  subsample_prop = 0.8,
  num_subsamples = 10,
  start_seed = 123,
  cluster_method
)
}
\arguments{
\item{data}{The dataset to be analyzed}

\item{n}{Number of observations to generate. Defaults to 100. Note: If n does not
split evenly by desired_k, it will be rounded up}

\item{min_k}{The starting number of clusters (k) for evaluation. Defaults to 2.}

\item{max_k}{The ending number of clusters (k) for evaluation. Defaults to 5.}

\item{subsample_prop}{Proportion of the dataset to use for each resample. Defaults to 0.8.}

\item{num_subsamples}{Number of resamples to generate. Defaults to 10.}

\item{start_seed}{Random seed for reproducibility. Defaults to 123.}

\item{cluster_method}{Specified clustering algorithm (i.e. "kmeans")}
}
\value{
A data frame where each row corresponds to a tested k cluster. The data frame
includes the similarity score, runtime, and a best_clust column, which is
blank for all rows except the one with the lowest similarity score, where it
is marked as "BEST".
}
\description{
Computes cluster score for clusters using k-Means to determine the optimal
number of clusters.
}
\examples{
data <- simulate_mvn(n = 10)
crab(data = data, max_k = 4, cluster_method = "kmeans")
}
