% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crab_sim_params.R
\name{crab_sim_params}
\alias{crab_sim_params}
\title{Choosing simulation parameters for CRAB Algorithm}
\usage{
crab_sim_params(
  data_info,
  min_k,
  max_k,
  num_subsamples,
  n = 100,
  variance,
  desired_k = 3,
  subsample_prop = 0.8,
  num_runs,
  cluster_method,
  start_seed = 123
)
}
\arguments{
\item{data_info}{A list containing the dataset, the original column names, and placeholder
NULL values for user-specified columns and parameters; the output from crab_prep()}

\item{min_k}{The starting number of clusters (k) for evaluation}

\item{max_k}{The ending number of clusters (k) for evaluation}

\item{num_subsamples}{Number of resamples to generate}

\item{n}{Number of observations to generate. Defaults to 100. Note: If n does not split
evenly by desired_k, it will be rounded up}

\item{desired_k}{The specific number of clusters (k) to evaluate. Defaults to 3.}

\item{subsample_prop}{Proportion of the dataset to use for each resample. Defaults to 0.8.}

\item{num_runs}{Number of iterations}

\item{cluster_method}{Specified clustering algorithm (i.e. "kmeans")}

\item{start_seed}{Reproducibility seed for running crab()}

\item{covariance}{List of variance values or single value used to construct the covariance
matrix for the dataset}
}
\value{
An updated list containing the data, its original column names, NULL for
user-specified columns (since the data is simulated), and the user-defined
simulation parameters.
}
\description{
Updates crab_prep() object allowing the user to specify simulation-specific
parameters to be used in crab().
}
\examples{
# must run simulated data first
simulated_data <- simulate_mvn(n = 10)

# create crab_setup with data, prep, and sim. params pipeline
crab_setup1 <- simulated_data |>
               crab_prep() |>
               crab_sim_params(min_k = 2, max_k = 3, num_subsamples = 10,
                               num_runs = 2, variance = 0.3,
                               cluster_method = "kmeans")
crab_setup1
}
