% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pointvec.R
\name{get_pointvec}
\alias{get_pointvec}
\title{Grab all pairwise comparisons for a given observation i and all other observations j in the mean matrix.}
\usage{
get_pointvec(i, mean_matrix_score)
}
\arguments{
\item{i}{Index of a point in dataset}

\item{mean_matrix_score}{The output score element produced by mean_matrix() (accessed by
mean_matrix$score).}
}
\value{
Generates a list of all of the pairwise comparisons between observations i and
every other observation j in the mean matrix.
}
\description{
Grab all pairwise comparisons for a given observation i and all other observations j in the mean matrix.
}
\examples{
# simulated data
mvn_df <- simulate_mvn(n = 400, variance = 0.4, desired_k = 3)

# mean matrix
mean_mat_mvn <- mean_matrix(data = mvn_df,
                            k = 3,
                            subsample_prop = 0.8,
                            num_subsamples = 10,
                            cluster_method = "kmeans")

# get point contributions
get_pointvec(i = 67, mean_matrix_score = mean_mat_mvn$score)
}
